requirements.txt
beautifulsoup4==4.13.4
faiss_cpu==1.11.0
langchain==0.3.25
langchain_community==0.3.24
langchain_core==0.3.63
langchain_openai==0.3.19
langchain_text_splitters==0.3.8
langgraph==0.4.8
python-dotenv==1.1.0
Requests==2.32.3
typing_extensions==4.14.0
